* commit '803dd78a6f8d7abcce170f984781f4ba18f72f96':
  Convert some junit integration tests to testng
  Clean up deprecated DrListTool uses
  Misc unused classes and warnings
  Remove unsafe warnings
  Statement unecessarry nested in else clause
  Trim DrMapTool
  Trim DrSetTool
  Remove DrSetTool.makeTreeSet
  Change junit import
  Remove DrGenericsFactory
  Remove unecessary suppress warnings
  Remove unecessary casts
  Remove some unused type parameters
  Repair generics in Nodes

+++ src/main/java/com/hotpads/datarouter/client/imp/hbase/cluster/DRHRegionInfo.java
@@ -2,7 +2,7 @@ package com.hotpads.datarouter.client.imp.hbase.cluster;
 
 import java.util.Random;
 
-import junit.framework.Assert;
+import org.junit.Assert;
 
 import org.apache.hadoop.hbase.HRegionInfo;
 import org.apache.hadoop.hbase.HServerLoad;
@@ -131,7 +131,7 @@ implements Comparable<DRHRegionInfo<?>>{
 		String totalKvsString = DrNumberFormatter.addCommas(totalKvs);
 		long compactingKvs = load.getCurrentCompactedKVs();
 		if(totalKvs==compactingKvs){ return totalKvsString; }
-		int percentCompacted = (int)((double)100 * (double)compactingKvs / (double)totalKvs);
+		int percentCompacted = (int)((double)100 * (double)compactingKvs / totalKvs);
 		return totalKvsString + " ["+percentCompacted+"%]";
 	}
