+++ src/main/java/com/hotpads/util/core/collections/arrays/LongArray.java
@@ -6,7 +6,7 @@ import java.util.LinkedList;
 import java.util.List;
 import java.util.ListIterator;
 
-import junit.framework.Assert;
+import org.junit.Assert;
 
 import org.junit.Test;
 
@@ -290,7 +290,7 @@ public class LongArray implements List<Long>, Comparable<List<Long>>{
 	@Override
 	public <T> T[] toArray(T[] wrapperArray) {
 		for(int i=0; i < this.size; ++i){
-			wrapperArray[i] = (T)(Long)(this.array[i]==NULL?null:this.array[i]);
+			wrapperArray[i] = (T)(this.array[i]==NULL?null:this.array[i]);
 		}
 		return wrapperArray;
 	}
